Add `zs:pop10` value for population candidate value
NYC (and other neighborhoods) don't appear to have `gn:population` values, so inputs like `Chelsea` are being punished in scoring for no other reason than lack of value (even the default `1` value isn't helping).  This fix uses the `zs:pop10` value for population when non-zero and `gn:population` is not available.  It's likely that `gn:population` and `zs:pop10` presence are mutually exclusive.