feat(download): Clarify downloader name in info logs
It's often the case that multiple Pelias importers are run in a single
terminal simultaneously (for example, the `pelias/docker` setup does
this).

With that in mind, be more explicit about which importer log lines are
coming from can be helpful. This adds `whosonfirst-` to some relevant
WOF download log lines.