fix(sqlite): remove obsolete index creation
The `obsolete` index creation is not compatible with the PIP service.

In the case where the sqlite databse does not yet have the `obsolete`
index, all PIP workers will attempt to create it, and most will fail to
obtain a write lock.

This causes the workers, and thus the whole PIP service to shut down,
without updating the sqlite db with the new index. Therefore no progress
can be made and the PIP service can never start.

It's likely this wasn't caught during development because the SQLite
database used for testing already had the index. However, the databases
downloaded from dist.whosonfirst.org do not _yet_ have it.

While this index is important for performance when reading from the
SQLite database, it will eventually make its way into the official
files.

In the meantime, the `dataHost` parameter in `pelias.json` can be used
in combination with a custom SQLite db with the index already created,
if max performance is desired.