feat(multi-lang): Add index for all wof languages (when differents to default language)
This importer did not index preferred names of all languages.
We need this for the multi-lang search in autocomplete.

I remove all elements when they are present in the default index. That means, if we say New York in the default language and New York in FRA, I will drop the FRA entry. This should prevent a massive increase of ES indexes.