chore(tests): Reduce overtesting of extractFields module
The tests for the extractFields module were repeatedly using full
objects for every test, which meant that functionality irrelevant for
individual unit tests was being tested.

This makes it much harder to change functionality, as a minor change to
one area will cause dozens of unrelated tests to fail.

Additionally, the test file is now much smaller and easier to read.