Refresh readme
Thanks to all the great recent changes, the README has drifted a bit.

This change is the result of a full read through, making things more
concise or accurate wherever possible.

In particular, I paid attention to new best practices as a result of our
work to support country-specific SQLite downloads. Users are now
encouraged to set `countryCode` properly to download only the data they
require.