fix(Dockerimage): Remove unnecessary compiler toolchain dependencies
After `better-sqlite3` added support for pre-compiled binaries in
https://github.com/JoshuaWise/better-sqlite3/releases/tag/v6.0.0, we no
longer need to install a compiler toolchain to run `npm install` in our
Docker images.

https://github.com/pelias/docker-baseimage/pull/23 is workong on
removing the compiler toolchain from our Pelias baseimages. In order for
the toolchain to be removed from the whosonfirst image in particular, we
also need to remove those dependencies here.

Until that PR is merged, this change is effectively a no-op. After,
between the two PRs we reduce the size of the whosonfirst docker image
from 490MB to 261MB, an impressive 221MB savings!